Move glem e2e test to the end

+++ tests/end2end-tests/graphstorm-nc/mgpu_test.sh
@@ -184,18 +184,6 @@ python3 -m graphstorm.run.gs_node_classification --workspace $GS_HOME/training_s
 
 error_and_exit ${PIPESTATUS[0]}
 
-echo "**************dataset: MovieLens classification, GLEM co-training, RGCN layer: 1, node feat: BERT nodes: movie, user inference: mini-batch save model save emb node"
-python3 -m graphstorm.run.gs_node_classification --workspace $GS_HOME/training_scripts/gsgnn_np/ --num-trainers $NUM_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_text_train_val_1p_4t/movie-lens-100k-text.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_nc_utext_glem.yml  --save-model-path /data/gsgnn_nc_ml_text/ --topk-model-to-save 1 --num-epochs 3 | tee train_log.txt
-
-error_and_exit ${PIPESTATUS[0]}
-
-cnt=$(ls -l /data/gsgnn_nc_ml_text/ | grep epoch | wc -l)
-if test $cnt != 2 # expect to save 2 * k models for GLEM
-then
-    echo "The number of save models $cnt is not equal to the specified topk 1 * 2"
-    exit -1
-fi
-
 best_epoch=$(grep "successfully save the model to" train_log.txt | tail -1 | tr -d '\n' | tail -c 1)
 echo "The best model is saved in epoch $best_epoch"
 
@@ -254,5 +242,17 @@ echo "**************dataset: MovieLens classification, RGCN layer: 1, node feat:
 python3 -m graphstorm.run.gs_node_classification --workspace $GS_HOME/training_scripts/gsgnn_np/ --num-trainers $NUM_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_text_train_val_1p_4t/movie-lens-100k-text.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_nc_utext.yaml  --save-model-path /data/gsgnn_nc_ml_text/ --topk-model-to-save 1 --save-embed-path /data/gsgnn_nc_ml_text/emb/ --num-epochs 3 --lm-train-nodes 0 | tee train_log.txt
 
 error_and_exit ${PIPESTATUS[0]}
+
+echo "**************dataset: MovieLens classification, GLEM co-training, RGCN layer: 1, node feat: BERT nodes: movie, user inference: mini-batch save model save emb node"
+python3 -m graphstorm.run.gs_node_classification --workspace $GS_HOME/training_scripts/gsgnn_np/ --num-trainers $NUM_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_text_train_val_1p_4t/movie-lens-100k-text.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_nc_utext_glem.yml  --save-model-path /data/gsgnn_nc_ml_text/ --topk-model-to-save 1 --num-epochs 3 | tee train_log.txt
+
+error_and_exit ${PIPESTATUS[0]}
+
+cnt=$(ls -l /data/gsgnn_nc_ml_text/ | grep epoch | wc -l)
+if test $cnt != 2 # expect to save 2 * k models for GLEM
+then
+    echo "The number of save models $cnt is not equal to the specified topk 1 * 2"
+    exit -1
+fi
 
 rm -fr /data/gsgnn_nc_ml_text/*
